backend/requirements.txt
fastapi
motor
passlib[bcrypt]
pydantic-settings
pymongo
python-jose[cryptography]
typer